Moving External API tests into separate phpunit group (task #5024)
