+++ tests/TestCase/Controller/Api/UsersControllerTest.php
@@ -183,83 +183,4 @@ class UsersControllerTest extends IntegrationTestCase
 
         $this->assertEquals($data['first_name'], $entity->get('first_name'));
     }
-
-    public function testExternalApiUsersCRUD()
-    {
-        $response = $this->apiClient->post('/api/users/token.json', [
-            'username' => getenv('DEV_USER'),
-            'password' => getenv('DEV_USER'),
-        ]);
-
-        $this->assertTrue($response->isOk(), "Couldn't fetch API token from users");
-
-        $responseBody = json_decode($response->body(), true);
-        unset($response);
-
-        $token = $responseBody['data']['token'];
-        $response = $this->sendAuthPost('/api/users/index.json', [], ['token' => $token]);
-        $this->assertTrue($response->isOk());
-
-        $usersResponse = json_decode($response->body(), true);
-
-        $this->assertTrue($usersResponse['success'], "Failed response on API users index call");
-        $this->assertNotEmpty($usersResponse['data'], "No users found in API /users/index.json");
-        unset($response);
-
-        $data = [
-            'username' => 'test-api-external',
-            'password' => 'test-api-external',
-            'active' => 1,
-            'is_superuser' => 0,
-        ];
-
-        $response = $this->sendAuthPost('/api/users/add.json', $data, ['token' => $token]);
-        $this->assertTrue($response->isOk());
-
-        $userCreated = json_decode($response->body(), true);
-        unset($response);
-
-        $response = $this->sendAuthPost(
-            '/api/users/view/' . $userCreated['data']['id'] . '.json',
-            [],
-            ['token' => $token]
-        );
-        $this->assertTrue($response->isOk());
-
-        $userView = json_decode($response->body(), true);
-
-        $this->assertEquals($userCreated['data']['id'], $userView['data']['id']);
-        unset($response);
-
-        $response = $this->sendAuthPost(
-            '/api/users/edit/' . $userView['data']['id'] . '.json',
-            ['is_superuser' => 1],
-            ['token' => $token]
-        );
-
-        $this->assertTrue($response->isOk());
-        unset($response);
-
-        $response = $this->sendAuthPost('/api/users/delete/' . $userView['data']['id'] . '.json', [], ['token' => $token]);
-
-        $this->assertTrue($response->isOk());
-        unset($response);
-
-        $response = $this->sendAuthPost('/api/users/view/' . $userView['data']['id'] . '.json', [], ['token' => $token]);
-        $this->assertFalse($response->isOk());
-    }
-
-    protected function sendAuthPost($url = '', $data = [], $headers = [])
-    {
-        $options = [
-            'headers' => [
-                'Authorization' => 'Bearer ' . $headers['token'],
-            ],
-            'type' => 'json',
-        ];
-
-        $response = $this->apiClient->post($url, $data, $options);
-
-        return $response;
-    }
 }
